add exist checking

+++ index.js
@@ -1,10 +1,10 @@
 #!/usr/bin/env node
 
-const shell = require('shelljs');
-const fs = require('fs');
-const path = require('path');
-const npm = require('./npm');
-const cdnjs = require('./cdnjs');
+const shell = require("shelljs");
+const fs = require("fs");
+const path = require("path");
+const npm = require("./npm");
+const cdnjs = require("./cdnjs");
 
 const log = console.log;
 const options = process.argv.slice(2);
@@ -23,7 +23,7 @@ function initFolder() {
 
 function installPkg() {
   if (options.length !== 3) {
-    log('Error on arguments');
+    log("Error on arguments");
     shell.exit(1);
   }
   const pkgName = options[1];
@@ -44,43 +44,45 @@ function clearPkg() {
   const currPath = shell.pwd().toString();
   const pkgFolderList = fs.readdirSync(currPath);
   pkgFolderList.forEach((pkgFolder) => {
-    const folderPath = path.join(currPath, pkgFolder, 'node_modules');
-    const folderList = fs.readdirSync(folderPath);
-    folderList.forEach((folder) => {
-      if (!stableVersion.test(folder)) {
-        fs.rmdirSync(path.join(folderPath, folder), { recursive: true });
-      }
-    })
-  })
+    const folderPath = path.join(currPath, pkgFolder, "node_modules");
+    if (fs.existsSync(folderPath)) {
+      const folderList = fs.readdirSync(folderPath);
+      folderList.forEach((folder) => {
+        if (!stableVersion.test(folder)) {
+          fs.rmdirSync(path.join(folderPath, folder), { recursive: true });
+        }
+      });
+    }
+  });
 }
 
 switch (operation) {
-  case 'init': {
+  case "init": {
     initFolder();
     break;
   }
-  case 'install': {
+  case "install": {
     installPkg();
     break;
   }
-  case 'past': {
+  case "past": {
     npm.pastVersion(options[1], parseInt(options[2], 10));
     break;
   }
-  case 'i': {
+  case "i": {
     initFolder();
     installPkg();
     break;
   }
-  case 'list': {
+  case "list": {
     // eslint-disable-next-line func-names
     (async function () {
       const pkgList = await cdnjs.getPkgList();
       log(pkgList);
-    }());
+    })();
     break;
   }
-  case 'sync': {
+  case "sync": {
     // eslint-disable-next-line func-names
     (async function () {
       const pkgList = await cdnjs.getPkgList();
@@ -89,14 +91,14 @@ switch (operation) {
       pkgList.forEach((pkg) => {
         shell.exec(`faststatic i ${pkg} ${lastCount}`);
       });
-    }());
+    })();
     break;
   }
-  case 'clean':
-    clearPkg(); 
+  case "clean":
+    clearPkg();
     break;
   default:
-    log('Error on arguments');
+    log("Error on arguments");
     shell.exit(-1);
     break;
 }
